Export the video model in fp16, which is the only form that fits
The twelve-frame window in fp32 exhausts 17 GiB at HD and takes twenty seconds.
In fp16 it takes 234 ms with 6.7 GiB to spare.

The demand is simply twelve times one frame's, and one frame already uses a
meaningful fraction of the card, so halving it was the move that mattered.
Everything tried before it is recorded in the doc and none of it worked:
enable_mem_pattern=False is a real 5x and still does not fit,
arena_extend_strategy adds nothing, and spatial chunking of the temporal blocks
adds nothing at that size either.

fp16 is also what iw3 runs these models at -- its inference path is under
torch.autocast throughout -- so the fp32 graph was the anomaly rather than this.
Accuracy against the fp32 PyTorch reference is max 2.4e-03, mean 9.1e-05 on an
output in 0..1, in a region the model is inventing pixels for.

One change was needed in the model to allow it. preprocess() called
mask.float(), which puts an fp32 tensor into an otherwise fp16 network and the
first convolution refuses it. iw3 writes .float() there meaning "make this bool
numeric"; following the input's dtype says the same thing, is identical for an
fp32 model, and lets the graph be exported in half. The golden tests still pass
at difference 0.

The image graph is left in fp32 for now. Halving it would take it from 35.4 ms
an eye to 18.3, but the plugin's buffers are fp32 and feeding a half graph needs
a cast either side, so it belongs with the plugin work rather than here.

The doc also carries a correction. The 6.7x and 26 ms figures quoted for
monobw_inpaint came from benchmarking PyTorch under autocast, which is fp16,
against an ONNX path that is fp32 -- not like for like. Per eye at HD the real
numbers are PyTorch fp16 12.0 ms, PyTorch fp32 23.1 ms, ONNX fp32 35.4 ms, so
the plugin costs about 71 ms a frame and roughly 15x row_flow_v3.

+++ stereo_inpaint.py
@@ -570,7 +570,12 @@ class LightInpaintV1(nn.Module):
         if closing:
             mask = _mask_closing(mask)
         else:
-            mask = mask.float()
+            # iw3 writes mask.float() here, meaning "make this bool numeric".
+            # Following the input's dtype says the same thing and is identical
+            # for an fp32 model, but it also lets the whole graph be exported in
+            # half precision -- a hard-coded .float() puts an fp32 tensor into
+            # an otherwise fp16 network and the first convolution refuses it.
+            mask = mask.to(x.dtype)
 
         mask = _dilate_inner(mask, n_iter=inner_dilation, base_width=base_width)
         mask = _dilate_outer(mask, n_iter=outer_dilation, base_width=base_width)
@@ -839,7 +844,12 @@ class LightVideoInpaintV1(nn.Module):
         if closing:
             mask = _mask_closing(mask)
         else:
-            mask = mask.float()
+            # iw3 writes mask.float() here, meaning "make this bool numeric".
+            # Following the input's dtype says the same thing and is identical
+            # for an fp32 model, but it also lets the whole graph be exported in
+            # half precision -- a hard-coded .float() puts an fp32 tensor into
+            # an otherwise fp16 network and the first convolution refuses it.
+            mask = mask.to(x.dtype)
 
         mask = _dilate_inner(mask, n_iter=inner_dilation, base_width=base_width)
         mask = _dilate_outer(mask, n_iter=outer_dilation, base_width=base_width)
